Removing the old _(){} functions from aliases and moving then to old-aliased/*.sh - [DEVELOP:{@hjunior}]

+++ bash_aliases.sh
@@ -12,15 +12,16 @@
 
 # inspiRED by: https://github.com/mathiasbynens/dotfiles
 
+# This is the only function in this file, intentionally
+
 # @function: Check if a command exists.
 # @param $1 [Req] : The command to check.
-__hhs_has() { 
-    type "$1" > /dev/null 2>&1
+__hhs_has() {
+    type "$1" >/dev/null 2>&1
 }
 
 # -----------------------------------------------------------------------------------
 # Navigational
-alias ..='function _() { [ -z "$1" ] && cd ..; if [ -n "$1" ]; then old_pwd=`pwd`; for x in `seq 1 "$1"`; do cd ..; done; fi; export OLDPWD="$old_pwd"; };_'
 alias ...='cd ../..'
 alias ....='cd ../../..'
 alias .....='cd ../../../..'
@@ -38,16 +39,13 @@ alias q="exit 0"
 alias sudo='sudo '
 
 # Always use color output for `ls`
-alias ls='command ls ${COLOR_FLAG} -F'
+alias ls='\ls ${COLOR_FLAG} -F'
 
 # List all files colorized in long format
 alias l='ls -lh'
 
 # List all directories
 alias lsd="ls -d */"
-
-# List all file names sorted by name
-alias lss='function _() { col=$1; [ -z "$1" ] && col=9; ls -la | sort -k "$col"; };_'
 
 # List all files colorized in long format, including dot files
 alias ll='ls -lah'
@@ -69,10 +67,11 @@ alias rm='rm -iv'
 alias cp='cp -iv'
 alias mv='mv -iv'
 
-# Setting some defaults
+# Nice command replacements
 alias cls='clear'
 alias df='df -H'
 alias du='du -hcd 1'
+alias psg='ps aux | grep -v grep | grep -i -e VSZ -e'
 
 # Use vim instead of vi
 __hhs_has "vim" && alias vi='vim'
@@ -84,9 +83,6 @@ alias less='less -R'
 # Make mount command output pretty and human readable format
 alias mount='mount | column -t'
 
-# Create all folders using a dot notation path and immediatelly change into it
-alias mkcd='function _() { if [ -n "$1" -a ! -d "$1" ]; then dir="${1//.//}"; mkdir -p $dir; pushd $dir >/dev/null; echo "${GREEN}Directory created: $dir ${NC}"; fi; };_'
-
 # Top shortcut ordered by cpu
 alias cpu='top -o cpu'
 
@@ -96,15 +92,6 @@ alias mem='top -o rsize'
 # Base64 encode shortcuts
 __hhs_has "base64" && alias encode="base64"
 
-# Linux boxes have a different syntaxes for some commands, so we craete the alias to match the correct OS.
-if [ "Linux" = "$MY_OS" ]; then 
-    alias ised="sed -i'' -r"
-    __hhs_has "base64" && alias decode='base64 -d'
-elif [ "Darwin" = "$MY_OS" ]; then
-    alias ised="sed -i '' -E"
-    __hhs_has "base64" && alias decode='base64 -D'
-fi
-
 # Date and time shortcuts
 alias week='date +%V'
 alias now='date +"(Week:%V) %Y-%m-%d %T %Z"'
@@ -113,30 +100,36 @@ alias ts='date "+%s%S"'
 # macOS has no `wget, so using curl instead`
 __hhs_has "wget" || alias wget='curl -O'
 
-# Generate a random number int the range <min> <max>
-alias rand='function _() { test -n "$1" -a -n "$2" && echo "$(( RANDOM % ($2 - $1 + 1 ) + $1 ))" || echo "Usage: rand <min> <max>"; };_'
-
 # Reload the bash session
 alias reload='cls; \. ~/.bashrc && echo -e "${HHS_WELCOME}"'
 
-# Kills all processes specified by $1
-alias pk='function _() { test -n "$1" && plist $1 kill; };_'
-
-# Swaps between PS1 & PS2 prompts 
+# Swaps between PS1 & PS2 prompts
 alias ps1='export PS1=$PS1_STYLE'
 alias ps2='export PS1=$PS2_STYLE'
 
 # -----------------------------------------------------------------------------------
-# Tool aliases
+# OS based aliases
 
-# Tree: List all directories recursively (Nth level depth) as a tree
-__hhs_has "tree" && alias lt='function _() { test -n "$1" -a -n "$2" && tree $1 -L $2 || tree $1; };_'
+# Linux boxes have a different syntaxes for some commands, so we craete the alias to match the correct OS.
+
+# -- LINUX --
+if [ "Linux" = "$MY_OS" ]; then
+    alias ised="sed -i'' -r"
+    __hhs_has "base64" && alias decode='base64 -d'
+# -- DARWIN --
+elif [ "Darwin" = "$MY_OS" ]; then
+    alias ised="sed -i '' -E"
+    __hhs_has "base64" && alias decode='base64 -D'
+fi
+
+# -----------------------------------------------------------------------------------
+# Tool aliases
 
 # Jenv: Set JAVA_HOME using jenv
 __hhs_has "jenv" && alias jenv_set_java_home='export JAVA_HOME="$HOME/.jenv/versions/`jenv version-name`"'
 
 # Dropbox: Recursively delete Dropbox conflicted files from the current directory
-[ -d "$DROPBOX" ] && alias db-clean="find . -name *\ \(*conflicted* -exec rm -v {} \;"
+[ -d "$DROPBOX" ] && alias db-cleanup="find . -name *\ \(*conflicted* -exec rm -v {} \;"
 
 # -----------------------------------------------------------------------------------
 # Python aliases
@@ -145,7 +138,7 @@ if __hhs_has "python"; then
 
     # linux has no `json_pp`, so using python instead
     __hhs_has "json_pp" || alias json_pp='python -m json.tool'
-    
+
     # Evaluate mathematical expression
     alias calc='python -c "import sys,math; print(eval(\" \".join(sys.argv[1:])));"'
 
@@ -156,8 +149,6 @@ if __hhs_has "python"; then
     # Generate a UUID
     alias uuid='python -c "import uuid as ul; print(ul.uuid4())"'
 
-    # Convert unicode to hexadecimal
-    alias utoh='function _() { print-uni.py $* | hexdump -Cb; };_'
 fi
 
 # -----------------------------------------------------------------------------------
@@ -192,10 +183,10 @@ __hhs_has "ifconfig" && alias ips="ifconfig -a | grep -o 'inet6\? \(addr:\)\?\s\
 # -----------------------------------------------------------------------------------
 # Mac Stuff
 
-if [ "Darwin" = "$MY_OS" ]; then 
+if [ "Darwin" = "$MY_OS" ]; then
 
     # Delete all .DS_store files
-    alias clean-ds="find . -type f -name '*.DS_Store' -ls -delete"
+    alias ds-cleanup="find . -type f -name '*.DS_Store' -ls -delete"
 
     # Flush Directory Service cache
     __hhs_has "dscacheutil" && alias flush="dscacheutil -flushcache && killall -HUP mDNSResponder"
@@ -252,10 +243,6 @@ alias enable-line-wrap='tput smam'
 # Disable line wrapping
 alias disable-line-wrap='tput rmam'
 
-# 
-# EXPERIMENTAL {
-# 
-
 # -----------------------------------------------------------------------------------
 # Git Stuff
 
@@ -272,12 +259,9 @@ if __hhs_has "git"; then
     alias __hhs_git_add='git add'
     alias __hhs_git_commit='git commit -m'
     alias __hhs_git_amend='git commit --amend --no-edit'
-    alias __hhs_git_branch_all='function _() { test -n "$1" -a -n "$2" && for x in $(find "$1" -maxdepth 1 -type d -iname "$2"); do cd $x; pwd; git status | head -n 1; cd - > /dev/null; done || echo "Usage: gba <dirname> <fileext>"; };_'
-    alias __hhs_git_status_all='function _() { test -n "$1" && for x in $(find "$1" -maxdepth 1 -type d -iname "*.git"); do cd $x; pwd; git status; cd - > /dev/null; done || echo "Usage: gsa <dirname>"; };_'
     alias __hhs_git_pull_rebase='git pull --rebase'
     alias __hhs_git_push='git push origin HEAD'
     alias __hhs_git_show='git diff-tree --no-commit-id --name-status -r'
-    alias __hhs_git_diff_show='function _() { git diff $1^1 $1 -- $2; };_'
     alias __hhs_git_difftool='git difftool -t opendiff'
 fi
 
@@ -286,8 +270,6 @@ fi
 
 if __hhs_has "gradle"; then
 
-    # Prefer using the wrapper instead of the command itself
-    alias __hhs_gradlew='function _() { [ -f "./gradlew" ] && ./gradlew $* || gradle $*; };_'
     alias __hhs_gradle_build='gw clean build'
     alias __hhs_gradle_run='gw bootRun -x Test'
     alias __hhs_gradle_test='gw Test'
@@ -302,26 +284,25 @@ fi
 # Docker stuff
 # inspiRED by https://hackernoon.com/handy-docker-aliases-4bd85089a3b8
 
-if __hhs_has "docker"; then
+if __hhs_has "docker" && docker info &> /dev/null; then
 
     alias __hhs_docker_images='docker images'
     alias __hhs_docker_service='docker service'
     alias __hhs_docker_logs='docker logs'
-    alias __hhs_docker_exec='function _() { [ -n "$2" ] && docker exec -it "$1" "$2" || docker exec -it "$1" /bin/sh; };_'
     alias __hhs_docker_remove='for next in $(docker volume ls -qf dangling=true); do echo "Removing Docker volume: $next"; docker volume rm $next; done'
     alias __hhs_docker_remove_image='docker rmi'
     alias __hhs_docker_ps='docker ps --format "{{.ID}} - {{.Names}} - {{.Status}} - {{.Image}}"'
-    alias __hhs_docker_pidof='function _() { docker ps | grep "$1" | awk '"'"'{print $1}'"'"'; };_'
-    alias __hhs_docker_tail_logs='function _() { docker logs -f $(docker ps | grep "$1" | awk '"'"'{print $1}'"'"'); };_'
     alias __hhs_docker_top='docker stats --format "table {{.Container}}\t{{.Name}}\t{{.CPUPerc}}\t{{.NetIO}}\t{{.BlockIO}}"'
     alias __hhs_docker_ls='docker container ls'
 fi
-
-#
-# } EXPERIMENTAL
-#
 
 # Source the custom alias shortcuts
 if [ -s "$HOME/.bash_aliasdef" ]; then
     \. "$HOME/.bash_aliasdef"
 fi
+
+# Load all functions that were previously aliased in here
+# shellcheck disable=SC2044
+for file in $(find "$HOME_SETUP/bin/old-aliased" -type f -name "*.sh" | sort); do
+    \. "$file"
+done
